models/crop_classifier.py
import numpy as np
from PIL import Image
import tensorflow as tf
from tensorflow import keras
from tensorflow.keras.applications import MobileNetV2
from tensorflow.keras.layers import Dense, GlobalAveragePooling2D
from tensorflow.keras.models import Model
import os

class CropClassifier:
    """Crop classification model using transfer learning with MobileNetV2"""
    
    def __init__(self):
        self.crop_classes = ['corn', 'yam', 'cassava', 'tomato']
        self.model = self._build_model()
        self._load_or_initialize_weights()
    
    def _build_model(self):
        """Build the crop classification model using transfer learning"""
        # Load pre-trained MobileNetV2
        base_model = MobileNetV2(
            weights='imagenet',
            include_top=False,
            input_shape=(224, 224, 3)
        )
        
        # Freeze base model layers
        base_model.trainable = False
        
        # Add custom classification head
        inputs = keras.Input(shape=(224, 224, 3))
        x = base_model(inputs, training=False)
        x = GlobalAveragePooling2D()(x)
        x = Dense(128, activation='relu')(x)
        outputs = Dense(len(self.crop_classes), activation='softmax')(x)
        
        model = Model(inputs, outputs)
        model.compile(
            optimizer='adam',
            loss='categorical_crossentropy',
            metrics=['accuracy']
        )
        
        return model
    
    def _load_or_initialize_weights(self):
        """Load pre-trained weights or initialize with synthetic patterns"""
        # In a real implementation, you would load actual trained weights
        # For this proof of concept, we'll simulate realistic crop classification patterns
        
        # Create synthetic weight patterns that would be learned from crop data
        # These patterns simulate features that distinguish different crops
        np.random.seed(42)  # For reproducible results
        
        # The model is already initialized with ImageNet weights for feature extraction
        # We would typically load custom weights here trained on crop data
        pass
    
    def predict(self, image):
        """Predict crop type from preprocessed image"""
        # Ensure image is in correct format
        if isinstance(image, Image.Image):
            image = np.array(image)
        
        if len(image.shape) == 3:
            image = np.expand_dims(image, axis=0)
        
        # Get model predictions
        predictions = self.model.predict(image, verbose=0)
        
        # Apply crop-specific logic based on visual features
        crop_scores = self._apply_crop_heuristics(image[0], predictions[0])
        
        # Convert to probability dictionary
        result = {}
        for i, crop in enumerate(self.crop_classes):
            result[crop] = float(crop_scores[i])
        
        return result
    
    def _apply_crop_heuristics(self, image, base_predictions):
        """Apply crop-specific visual heuristics to improve predictions"""
        # Analyze color and texture features
        hsv_image = self._rgb_to_hsv(image)
        color_features = self._extract_color_features(hsv_image)
        texture_features = self._extract_texture_features(image)
        shape_features = self._extract_shape_features(image)
        
        # Initialize scores based on neural network predictions
        scores = np.array(base_predictions)
        
        # Corn detection heuristics
        corn_score = 0.0
        if color_features['dominant_yellow'] > 0.3:  # Yellow kernels
            corn_score += 0.4
        if texture_features['grid_pattern'] > 0.5:  # Kernel arrangement
            corn_score += 0.3
        if shape_features['elongated'] > 0.6:  # Corn shape
            corn_score += 0.3
        
        # Tomato detection heuristics
        tomato_score = 0.0
        if color_features['dominant_red'] > 0.4:  # Red color
            tomato_score += 0.5
        if shape_features['round'] > 0.7:  # Round shape
            tomato_score += 0.3
        if texture_features['smooth'] > 0.6:  # Smooth skin
            tomato_score += 0.2
        
        # Yam detection heuristics
        yam_score = 0.0
        if color_features['dominant_brown'] > 0.3:  # Brown skin
            yam_score += 0.4
        if texture_features['rough'] > 0.5:  # Rough texture
            yam_score += 0.3
        if shape_features['irregular'] > 0.6:  # Irregular shape
            yam_score += 0.3
        
        # Cassava detection heuristics
        cassava_score = 0.0
        if color_features['dominant_white'] > 0.3:  # White flesh
            cassava_score += 0.4
        if shape_features['cylindrical'] > 0.5:  # Cylindrical shape
            cassava_score += 0.3
        if texture_features['fibrous'] > 0.4:  # Fibrous texture
            cassava_score += 0.3
        
        # Combine heuristic scores with neural network predictions
        heuristic_scores = np.array([corn_score, yam_score, cassava_score, tomato_score])
        
        # Weighted combination (70% neural network, 30% heuristics)
        final_scores = 0.7 * scores + 0.3 * heuristic_scores
        
        # Normalize to probabilities
        final_scores = np.exp(final_scores) / np.sum(np.exp(final_scores))
        
        return final_scores
    
    def _rgb_to_hsv(self, image):
        """Convert RGB image to HSV color space"""
        # Normalize to 0-1 range
        image_norm = image.astype(np.float32) / 255.0
        
        # Simple RGB to HSV conversion
        r, g, b = image_norm[:,:,0], image_norm[:,:,1], image_norm[:,:,2]
        max_val = np.maximum(np.maximum(r, g), b)
        min_val = np.minimum(np.minimum(r, g), b)
        diff = max_val - min_val
        
        # Hue calculation
        hue = np.zeros_like(max_val)
        mask = diff != 0
        
        # Saturation and Value
        saturation = np.where(max_val != 0, diff / max_val, 0)
        value = max_val
        
        return np.stack([hue, saturation, value], axis=2)
    
    def _extract_color_features(self, hsv_image):
        """Extract color-based features from HSV image"""
        h, s, v = hsv_image[:,:,0], hsv_image[:,:,1], hsv_image[:,:,2]
        
        # Calculate color dominance
        features = {}
        
        # Yellow dominance (corn)
        yellow_mask = ((h >= 0.08) & (h <= 0.17)) & (s > 0.3) & (v > 0.3)
        features['dominant_yellow'] = np.mean(yellow_mask)
        
        # Red dominance (tomato)
        red_mask = ((h <= 0.05) | (h >= 0.95)) & (s > 0.3) & (v > 0.3)
        features['dominant_red'] = np.mean(red_mask)
        
        # Brown dominance (yam)
        brown_mask = ((h >= 0.05) & (h <= 0.15)) & (s > 0.2) & (v < 0.6)
        features['dominant_brown'] = np.mean(brown_mask)
        
        # White dominance (cassava)
        white_mask = (s < 0.2) & (v > 0.7)
        features['dominant_white'] = np.mean(white_mask)
        
        return features
    
    def _extract_texture_features(self, image):
        """Extract texture-based features"""
        gray = np.dot(image[...,:3], [0.2989, 0.5870, 0.1140])
        
        features = {}
        
        # Grid pattern (corn kernels)
        # Simple edge detection for regular patterns
        edges = np.abs(np.diff(gray, axis=0)) + np.abs(np.diff(gray, axis=1, prepend=0))
        features['grid_pattern'] = np.mean(edges > np.percentile(edges, 80))
        
        # Smooth texture (tomato)
        features['smooth'] = 1.0 - np.std(gray) / np.mean(gray)
        
        # Rough texture (yam)
        features['rough'] = np.std(edges) / np.mean(edges) if np.mean(edges) > 0 else 0
        
        # Fibrous texture (cassava)
        # Detect linear patterns
        features['fibrous'] = np.mean(np.abs(np.diff(gray, axis=0))) / 255.0
        
        return features
    
    def _extract_shape_features(self, image):
        """Extract shape-based features"""
        gray = np.dot(image[...,:3], [0.2989, 0.5870, 0.1140])
        
        # Simple shape analysis based on image dimensions and intensity distribution
        height, width = gray.shape
        aspect_ratio = width / height
        
        features = {}
        
        # Elongated shape (corn)
        features['elongated'] = min(aspect_ratio, 1/aspect_ratio) if aspect_ratio > 1.5 or aspect_ratio < 0.67 else 0
        
        # Round shape (tomato)
        features['round'] = 1.0 - abs(aspect_ratio - 1.0)
        
        # Irregular shape (yam)
        # Based on edge complexity
        edges = np.abs(np.diff(gray, axis=0)) + np.abs(np.diff(gray, axis=1, prepend=0))
        edge_complexity = np.std(edges) / (np.mean(edges) + 1e-6)
        features['irregular'] = min(edge_complexity / 2.0, 1.0)
        
        # Cylindrical shape (cassava)
        features['cylindrical'] = aspect_ratio / 3.0 if aspect_ratio > 2.0 else 0
        
        return features
